Fix CI build failures and add Milestone 3: Export & Share, wired into the real capture flow
## CI failures from the previous push -- root cause and fixes

All three failing jobs (Android build/lint/tests, shared core+UI Android
build, shared core+UI iOS build) traced back to the same handful of real
compiler errors, all in code from the previous commit:

- shared-core/src/commonTest/.../edit/PageEditTest.kt:108 -- `0x80_00_00_FF`
  has its high bit set, so Kotlin infers it as Long, not Int (unlike Java,
  Kotlin types hex literals by unsigned magnitude, not by whether the bit
  pattern fits in 32 bits). Every other hex literal in that file already
  had `.toInt()`; this one was missed. Fixed.
- shared-ui/.../ui/CropEnhanceScreens.kt -- `ScreenColumn(modifier = modifier)`
  at two call sites: `ScreenColumn` has no `modifier` parameter. Removed the
  argument and the now-unused `modifier` parameter from both
  CropPageScreen/EnhancePageScreen (matches every sibling screen in
  ToollyApp.kt, none of which expose an unused modifier either).
- Same file -- missing imports for `Res.string.product_name` and
  `Res.string.rotate` (both resources exist; the import lines were missing).

Re-verified by hand this time with three systematic sweeps across every
shared-ui file rather than spot-reading: (1) every `Res.string/plurals/
drawable.X` reference cross-checked against having a matching import line,
(2) the project's own hardcoded-text regex
(scripts/validate_platform_parity.py's KOTLIN_USER_TEXT_PATTERNS) grepped
by hand since Python isn't available in this shell, (3) every hex literal
8+ digits long checked for `.toInt()`. Also caught and fixed the identical
`ScreenColumn(modifier = modifier)` mistake in the not-yet-pushed
ExportScreens.kt before it could fail CI a second time.

## Milestone 3 (partial, scope-corrected): Export & Share

Investigated the domain model (spike-capture/.../domain/model/Documents.kt)
before building Library filters or search, per this session's established
practice of checking real backing capability before building UI on top of
it. DocumentSummary has no name/title/category field -- only id, page
count, timestamps and lifecycle. The wireframes' "Receipts/IDs" filter
chips and title search both need a document-naming feature that doesn't
exist, and inventing non-functional filter/search UI over data that isn't
there would be exactly the kind of fake control this project's own
policies (and Milestone 2's ML-Kit finding) argue against. Library filters
and real search are dropped from this milestone rather than faked;
whoever picks this up next needs a real naming/tagging feature first
(touches the encrypted vault's persisted schema -- its own scoped,
security-reviewed piece of work, not a quick add).

Export & Share, by contrast, already has a complete, physically-validated
backend (AndroidDocumentExporter, wired via CaptureSpikeActivity's
onExportDocument) -- it just exposed it through a bare 4-button grid. This
commit adds the shared, wireframe-matching selection UI and wires it in
place of that grid, without touching the export mechanism at all:

- shared-ui/.../ui/ExportScreens.kt: `ExportBuilderScreen` (wireframe 5.1 --
  format chips, a premium-locked "Searchable PDF" affordance per
  docs/product/ENTITLEMENTS.md's presentation-only-until-billing precedent,
  optional preview slot, optional back) and `ExportPrivacyCheckScreen`
  (wireframe 5.3 -- explicit disclosure before a document leaves the
  encrypted vault, per docs/security/PRIVACY_READINESS.md, with Save/Share
  as the two exits). `ToollyExportFormat` is a shared-ui-local PDF/JPEG
  enum -- deliberately not the wireframe's "Long image" (no exporter
  support) and no quality tiers (nothing to compute a real size estimate
  from) -- shared-ui can't reference spike-capture's domain
  DocumentExportFormat directly anyway (wrong dependency direction), so
  the platform host maps between them.
- Wireframe 5.2 (Export preview) and 5.4 (Share options) are intentionally
  not separate screens: there's no real PDF-preview-render capability to
  show (ExportBuilderScreen's previewContent slot reuses the existing,
  working DocumentPageGrid instead of faking a preview), and Share already
  is the OS share sheet per architecture rule ES-12 -- a dedicated "share
  options" screen would just be a redundant tap in front of it.
- spike-capture/.../ui/CaptureSpikeScreen.kt: added `AppScreen.ExportBuilder`
  and `AppScreen.ExportPrivacyCheck` states between the existing Document
  and Library states; DocumentScreen's 4-button grid is now one "Export"
  button. onExportDocument's signature, CaptureSpikeActivity's composition
  root, and the actual PDF/JPEG-writing code are all completely untouched.

## Verification

:shared-ui and :shared-core commonMain metadata compilation cannot be
checked locally past what's already documented in the Milestone 1/2
commits (Windows verification-metadata.xml gap). This round, verification
leaned entirely on the systematic hand-checks described above plus reading
the actual CI failure logs line-by-line via `gh api .../logs
--allow-escape-sequences` (gh's --log-failed truncates before the compiler
diagnostics) rather than guessing. Every fix here maps to a specific
compiler error message from the previous run.

Refs #52

+++ shared-core/src/commonTest/kotlin/com/toolly/shared/edit/PageEditTest.kt
@@ -104,7 +104,7 @@ class PageEditTest {
 
     @Test
     fun colorAdjustPreservesAlpha() {
-        val translucentBlue = 0x80_00_00_FF
+        val translucentBlue = 0x80_00_00_FF.toInt()
         val source = PixelBuffer.solid(1, 1, translucentBlue)
 
         val result = ColorAdjust.apply(source, EnhancementMode.CLEAN, intensity = 0.5f)
